+++ packages/tswebextension/src/lib/common/companies-db-service/trackers-min.ts
@@ -1,5 +1,5 @@
 export const rawCompaniesDb = {
-	"timeUpdated": "2025-08-25T10:05:23.938Z",
+	"timeUpdated": "2025-09-29T10:05:24.871Z",
 	"categories": {
 		"0": "audio_video_player",
 		"1": "comments",