+++ packages/tswebextension/vitest.config.ts
@@ -1,0 +1,67 @@
+import { defineConfig, defineProject, type UserWorkspaceConfig } from 'vitest/config';
+
+import { ManifestVersionEnv } from './tasks/constants';
+
+/**
+ * Creates a test configuration for a specific manifest version.
+ *
+ * @param manifestVersion The manifest version to create the test for.
+ *
+ * @returns The test configuration.
+ */
+const createProjectForManifestVersion = (
+    manifestVersion: ManifestVersionEnv,
+): UserWorkspaceConfig => defineProject({
+    test: {
+        name: `mv${manifestVersion}`,
+        env: {
+            MANIFEST_ENV: manifestVersion,
+        },
+        setupFiles: [
+            'fake-indexeddb/auto',
+            './vitest.setup.ts',
+        ],
+        environment: 'jsdom',
+        environmentOptions: {},
+        exclude: [
+            // node_modules are excluded by default and when we extend
+            // the default config, we need to exclude them explicitly.
+            '**/node_modules/**',
+            // eslint-disable-next-line max-len
+            `**/test/lib/mv${manifestVersion === ManifestVersionEnv.Second ? ManifestVersionEnv.Third : ManifestVersionEnv.Second}/**`,
+            '**/test/lib/common/**',
+        ],
+    },
+});
+
+export default defineConfig({
+    test: {
+        setupFiles: [
+            'fake-indexeddb/auto',
+            './vitest.setup.ts',
+        ],
+        environment: 'jsdom',
+        environmentOptions: {},
+        projects: [
+            defineProject({
+                test: {
+                    name: 'common',
+                    setupFiles: [
+                        'fake-indexeddb/auto',
+                        './vitest.setup.ts',
+                    ],
+                    environment: 'jsdom',
+                    environmentOptions: {},
+                    exclude: [
+                        // node_modules are excluded by default and when we extend
+                        // the default config, we need to exclude them explicitly.
+                        '**/node_modules/**',
+                        '**/test/lib/mv*/**',
+                    ],
+                },
+            }),
+            createProjectForManifestVersion(ManifestVersionEnv.Second),
+            createProjectForManifestVersion(ManifestVersionEnv.Third),
+        ],
+    },
+});